+++ CHANGELOG.md
@@ -7,6 +7,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ## [unreleased]
 ### Changed
 - Clearer error message when a loqusdb query fails for an instance that initially connected
+### Fixed
+- Safer way to update variant genes and compounds that avoids saving temporary decorators into variants' database documents
 
 ## [4.83]
 ### Added